updated Readme and app

+++ app.py
@@ -1,30 +1,38 @@
-
 from flask import Flask, jsonify, request
 import json
+import os
 
 app = Flask(__name__)
 
-FILE = 'notes.json'
+FILE = "notes.json"
 
 
+# Create file if it doesn't exist
+if not os.path.exists(FILE):
+    with open(FILE, "w") as f:
+        json.dump([], f)
+
+
+# Read notes
 def readnote():
-    with open(FILE, 'r') as f:
+    with open(FILE, "r") as f:
         return json.load(f)
 
 
+# Write notes
 def writenote(notes):
-    with open(FILE, 'w') as f:
+    with open(FILE, "w") as f:
         json.dump(notes, f, indent=4)
 
 
-# READ
+# GET all notes
 @app.route('/notes', methods=['GET'])
 def getnotes():
     notes = readnote()
     return jsonify(notes)
 
 
-# CREATE
+# POST create new note
 @app.route('/notes', methods=['POST'])
 def addnote():
     notes = readnote()
@@ -32,7 +40,7 @@ def addnote():
 
     new_note = {
         "id": len(notes) + 1,
-        "note": data["note"]
+        "note": data.get("note")
     }
 
     notes.append(new_note)
@@ -41,34 +49,37 @@ def addnote():
     return jsonify(new_note)
 
 
-# UPDATE
-@app.route('/notes/<int:id>', methods=['PUT'])
-def updatenote(id):
+@app.route('/notes', methods=['PUT'])
+def updatenote():
     notes = readnote()
     data = request.get_json()
 
+    note_id = data.get("id")
+    new_note = data.get("note")
+
     for note in notes:
-        if note["id"] == id:
-            note["note"] = data["note"]
+        if note["id"] == note_id:
+            note["note"] = new_note
             writenote(notes)
             return jsonify(note)
 
-    return {"message": "Note not found"}
+    return jsonify({"message": "Note not found"})
 
-
-# DELETE
-@app.route('/notes/<int:id>', methods=['DELETE'])
-def deletenote(id):
+@app.route('/notes', methods=['DELETE'])
+def deletenote():
     notes = readnote()
+    data = request.get_json()
+
+    note_id = data.get("id")
 
     for note in notes:
-        if note["id"] == id:
+        if note["id"] == note_id:
             notes.remove(note)
             writenote(notes)
-            return {"message": "Note deleted"}
+            return jsonify({"message": "Note deleted"})
 
-    return {"message": "Note not found"}
+    return jsonify({"message": "Note not found"})
+    
 
-
-if __name__ == '__main__':
+if __name__ == "__main__":
     app.run(debug=True)
